# spec/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Business Watchdog core API
  version: 1.0.0
  description: |
    Core contracts for P0 and P1. This document is a specification, not a running API.
    HMAC canonicalization and transport limits are normative in TZ sections 11 and 27.
    UI CRUD, auth, billing, export and credential rotation must be added before their implementation.
    All monetary amounts are decimal strings in minor units. All timestamps are RFC3339 UTC.
servers:
  - url: https://watchdog.example.invalid
...
paths:
...
  /api/v1/tenants/context:
    get:
      tags: [Tenancy]
      operationId: tenantContext
      security:
        - {UserSession: []}
      responses:
        "200":
          description: Resolved tenant context
          content:
            application/json:
              schema:
                type: object
                properties:
                  active_tenant_id: {type: [string, "null"], format: uuid}
                  tenant_context_id: {type: [string, "null"], format: uuid}
                  tenant_context_source: {type: [string, "null"]}
        "401": {$ref: "#/components/responses/Problem"}
...
  /api/v1/stores/{store_id}/verification/check:
    post:
      tags: [Stores]
      operationId: checkStoreVerification
      security:
        - {UserSession: []}
      description: SSRF-safe DNS/HTTP check; rate-limited.
      parameters:
        - name: store_id
          in: path
          required: true
          schema: {type: string, format: uuid}
        - {$ref: "#/components/parameters/CsrfToken"}
      responses:
        "200":
          description: Checked now
          content:
            application/json:
              schema: {$ref: "#/components/schemas/StoreVerification"}
        "401": {$ref: "#/components/responses/Problem"}
        "403": {$ref: "#/components/responses/Problem"}
        "404": {$ref: "#/components/responses/Problem"}
        "429": {$ref: "#/components/responses/Problem"}
...
  /api/v1/stores/{store_id}/scenarios:
    get:
      tags: [Checks]
      operationId: listCheckScenarios
      security:
        - {UserSession: []}
      parameters:
        - name: store_id
          in: path
          required: true
          schema: {type: string, format: uuid}
      responses:
        "200":
          description: Scenarios of the store (one payment_form scenario in P0)
          content:
            application/json:
              schema:
                type: object
                required: [data, next_cursor]
                properties:
                  data:
                    type: array
                    items: {$ref: "#/components/schemas/CheckScenario"}
                  next_cursor: {type: [string, "null"]}
        "401": {$ref: "#/components/responses/Problem"}
        "404": {$ref: "#/components/responses/Problem"}
    post:
      tags: [Checks]
      operationId: createCheckScenario
      security:
        - {UserSession: []}
      description: "Owner/admin. Deviation from the catalogue until the plugin manifest exists: the product URL is given directly (HTTPS, store origin). 409 scenario_exists."
      parameters:
        - name: store_id
          in: path
          required: true
          schema: {type: string, format: uuid}
        - {$ref: "#/components/parameters/CsrfToken"}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - {$ref: "#/components/schemas/CheckScenarioInput"}
                - {required: [product_url]}
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: {$ref: "#/components/schemas/CheckScenario"}
          headers:
            ETag:
              description: Quoted version for If-Match
              schema: {type: string}
        "401": {$ref: "#/components/responses/Problem"}
        "403": {$ref: "#/components/responses/Problem"}
        "404": {$ref: "#/components/responses/Problem"}
        "409": {$ref: "#/components/responses/Problem"}
        "422": {$ref: "#/components/responses/Problem"}
  /api/v1/scenarios/{scenario_id}:
    get:
      tags: [Checks]
      operationId: getCheckScenario
      security:
        - {UserSession: []}
      parameters:
        - name: scenario_id
          in: path
          required: true
          schema: {type: string, format: uuid}
      responses:
        "200":
          description: Scenario
          content:
            application/json:
              schema: {$ref: "#/components/schemas/CheckScenario"}
          headers:
            ETag:
              description: Quoted version for If-Match
              schema: {type: string}
        "401": {$ref: "#/components/responses/Problem"}
        "404": {$ref: "#/components/responses/Problem"}
    patch:
      tags: [Checks]
      operationId: updateCheckScenario
      security:
        - {UserSession: []}
      description: Owner/admin; If-Match carries the scenario version.
      parameters:
        - name: scenario_id
          in: path
          required: true
          schema: {type: string, format: uuid}
        - {$ref: "#/components/parameters/IfMatch"}
        - {$ref: "#/components/parameters/CsrfToken"}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/CheckScenarioInput"}
      responses:
        "200":
          description: New scenario version
          content:
            application/json:
              schema: {$ref: "#/components/schemas/CheckScenario"}
          headers:
            ETag:
              description: Quoted version for If-Match
              schema: {type: string}
        "401": {$ref: "#/components/responses/Problem"}
        "403": {$ref: "#/components/responses/Problem"}
        "404": {$ref: "#/components/responses/Problem"}
        "409": {$ref: "#/components/responses/Problem"}
        "422": {$ref: "#/components/responses/Problem"}
        "428": {$ref: "#/components/responses/Problem"}
  /api/v1/stores/{store_id}/checks:
    get:
      tags: [Checks]
      operationId: listCheckRuns
      security:
        - {UserSession: []}
      parameters:
        - name: store_id
          in: path
          required: true
          schema: {type: string, format: uuid}
        - name: limit
          in: query
          schema: {type: integer, minimum: 1, maximum: 100}
        - name: cursor
          in: query
          schema: {type: string, maxLength: 2048}
      responses:
        "200":
          description: "Runs, newest first"
          content:
            application/json:
              schema:
                type: object
                required: [data, next_cursor]
                properties:
                  data:
                    type: array
                    items: {$ref: "#/components/schemas/CheckRun"}
                  next_cursor: {type: [string, "null"]}
        "401": {$ref: "#/components/responses/Problem"}
        "404": {$ref: "#/components/responses/Problem"}
    post:
      tags: [Checks]
      operationId: createManualCheck
      security:
        - {UserSession: []}
      description: Owner/admin; one manual run per minute per store (429); 409 while a run is active; 422 if the store is not eligible.
      parameters:
        - name: store_id
          in: path
          required: true
          schema: {type: string, format: uuid}
        - {$ref: "#/components/parameters/IdempotencyKey"}
        - {$ref: "#/components/parameters/CsrfToken"}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [scenario_id]
              properties:
                scenario_id: {type: string, format: uuid}
      responses:
        "202":
          description: Durable logical run created
          content:
            application/json:
              schema:
                allOf:
                  - type: object
                    required: [run_id]
                    properties:
                      run_id: {type: string, format: uuid}
                  - {$ref: "#/components/schemas/CheckRun"}
        "401": {$ref: "#/components/responses/Problem"}
        "403": {$ref: "#/components/responses/Problem"}
        "404": {$ref: "#/components/responses/Problem"}
        "409": {$ref: "#/components/responses/Problem"}
        "422": {$ref: "#/components/responses/Problem"}
        "429": {$ref: "#/components/responses/Problem"}
  /api/v1/checks/{run_id}:
    get:
      tags: [Checks]
      operationId: getCheckRun
      security:
        - {UserSession: []}
      parameters:
        - name: run_id
          in: path
          required: true
          schema: {type: string, format: uuid}
      responses:
        "200":
          description: "Run with attempts, sanitized steps and ready artifacts"
          content:
            application/json:
              schema: {$ref: "#/components/schemas/CheckRunDetail"}
        "401": {$ref: "#/components/responses/Problem"}
        "404": {$ref: "#/components/responses/Problem"}
...
components:
  securitySchemes:
    IntegrationKeyId:
      type: apiKey
      in: header
      name: X-BW-Key-Id
      description: Key ID plus mandatory raw-body HMAC signature headers; key ID alone never authenticates.
    UserSession:
      type: apiKey
      in: cookie
      name: watchdog_session
      description: Same-origin Sanctum session with CSRF on mutations.
    WorkerToken:
      type: http
      scheme: bearer
      description: Scoped worker credential restricted to private ingress.
  parameters:
...
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: Strong ETag containing quoted incident revision, for example "3".
      schema: {type: string, pattern: '^"[0-9]+"$'}
    CsrfToken:
      name: X-XSRF-TOKEN
      in: header
      required: true
      schema: {type: string}
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: {type: string, minLength: 16, maxLength: 128}
  responses:
    Problem:
      description: Structured safe problem response; HTTP status determines retry policy.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
...
    Problem:
      type: object
      additionalProperties: false
      required: [code, message, request_id]
      properties:
        code: {type: string}
        message: {type: string}
        request_id: {type: string, format: uuid}
        errors: {type: array, items: {type: object}}
        retry_after_seconds: {type: integer, minimum: 0}
...
    StoreVerification:
      type: object
      required: [id, store_id, method, state]
      properties:
        id: {type: string, format: uuid}
        store_id: {type: string, format: uuid}
        method: {enum: [dns, plugin_challenge]}
        state: {enum: [pending, verified, failed, expired]}
        verified_origin: {type: [string, "null"]}
        expires_at: {type: [string, "null"], format: date-time}
        verified_at: {type: [string, "null"], format: date-time}
        reason_code: {type: [string, "null"]}
        attempts: {type: integer}
        last_checked_at: {type: [string, "null"], format: date-time}
...
    CheckScenario:
      type: object
      required:
        - id
        - store_id
        - name
        - mode
        - version
        - enabled
        - adapter_version
        - interval_seconds
        - supported_steps
        - untested_components
      properties:
        id: {type: string, format: uuid}
        store_id: {type: string, format: uuid}
        name: {type: string}
        mode: {const: payment_form}
        version: {type: integer, minimum: 1}
        enabled: {type: boolean}
        adapter_version: {type: string}
        product_external_id: {type: [string, "null"]}
        product_url: {type: [string, "null"], format: uri}
        cart_url: {type: [string, "null"], format: uri}
        checkout_url: {type: [string, "null"], format: uri}
        extra_allowed_origins:
          type: array
          items: {type: string, format: uri}
        synthetic_location:
          type: [object, "null"]
          properties:
            country: {type: string, pattern: "^[A-Z]{2}$"}
            postcode: {type: [string, "null"]}
        interval_seconds: {type: integer, minimum: 300, maximum: 86400}
        next_due_at: {type: [string, "null"], format: date-time}
        supported_steps:
          type: array
          items: {type: string}
        untested_components:
          type: array
          items: {type: string}
    CheckScenarioInput:
      type: object
      additionalProperties: false
      properties:
        name: {type: string, minLength: 1, maxLength: 100}
        product_url: {type: string, format: uri, description: HTTPS on the store origin.}
        cart_url: {type: [string, "null"], format: uri}
        checkout_url: {type: [string, "null"], format: uri}
        product_external_id: {type: [string, "null"], maxLength: 255}
        interval_seconds: {type: integer, minimum: 300, maximum: 86400}
        enabled: {type: boolean, description: Requires a verified store.}
        extra_allowed_origins:
          type: array
          maxItems: 10
          items: {type: string, format: uri}
          description: Public HTTPS gateway/CDN origins the check may load.
        synthetic_location:
          type: [object, "null"]
          additionalProperties: false
          required: [country]
          properties:
            country: {type: string, pattern: "^[A-Za-z]{2}$"}
            postcode: {type: [string, "null"], maxLength: 16}
    CheckRun:
      type: object
      required: [id, store_id, scenario_id, scenario_version, trigger, status, scheduled_at]
      properties:
        id: {type: string, format: uuid}
        store_id: {type: string, format: uuid}
        scenario_id: {type: string, format: uuid}
        scenario_version: {type: integer}
        trigger: {enum: [scheduled, manual, incident]}
        status: {enum: [queued, running, passed, failed, inconclusive, blocked, unsupported, cancelled]}
        error_code: {type: [string, "null"]}
        scheduled_at: {type: string, format: date-time}
        started_at: {type: [string, "null"], format: date-time}
        finished_at: {type: [string, "null"], format: date-time}
    CheckArtifact:
      type: object
      required: [id, kind, content_type, size_bytes, expires_at]
      properties:
        id: {type: string, format: uuid}
        kind: {enum: [screenshot]}
        content_type: {enum: [image/jpeg, image/png]}
        size_bytes: {type: integer}
        redaction_version: {type: string}
        expires_at: {type: string, format: date-time}
    CheckRunDetail:
      allOf:
        - {$ref: "#/components/schemas/CheckRun"}
        - type: object
          required: [attempts]
          properties:
            attempts:
              type: array
              items:
                type: object
                required: [id, attempt_number, status, steps, artifacts]
                properties:
                  id: {type: string, format: uuid}
                  attempt_number: {type: integer}
                  status: {enum: [running, passed, failed, inconclusive, blocked, unsupported, cancelled, expired]}
                  error_code: {type: [string, "null"]}
                  location: {type: string}
                  browser_version: {type: string}
                  started_at: {type: string, format: date-time}
                  finished_at: {type: [string, "null"], format: date-time}
                  diagnostics: {type: [object, "null"], description: redaction_version and relevant_errors (sanitized metadata).}
                  artifacts:
                    type: array
                    items: {$ref: "#/components/schemas/CheckArtifact"}
                  steps:
                    type: array
                    items:
                      type: object
                      required: [index, code, status]
                      properties:
                        index: {type: integer}
                        code: {type: string}
                        status: {enum: [passed, failed, skipped, inconclusive, blocked]}
                        error_code: {type: [string, "null"]}
                        started_at: {type: string, format: date-time}
                        finished_at: {type: string, format: date-time}
                        assertions:
                          type: array
                          items: {type: object}
                        network_summary:
                          type: array
                          items: {type: object}
...